+++ test/foundry/integration/RepayController.t.sol
@@ -235,7 +235,7 @@ contract RepayAdapterIntegrationTest is Test, CloberMarketSwapCallbackReceiver, 
         address borrowToken,
         uint256 collateralAmount,
         uint256 borrowAmount,
-        uint8 loanEpochs
+        uint16 loanEpochs
     ) internal returns (uint256 positionId) {
         positionId = loanPositionManager.nextId();
         IController.ERC20PermitParams memory permitParams = _buildERC20PermitParams(
